+++ core/trino-main/src/test/java/io/trino/operator/aggregation/TestDoubleMinAggregation.java
@@ -30,7 +30,7 @@ public class TestDoubleMinAggregation
     {
         BlockBuilder blockBuilder = DOUBLE.createBlockBuilder(null, length);
         for (int i = start; i < start + length; i++) {
-            DOUBLE.writeDouble(blockBuilder, (double) i);
+            DOUBLE.writeDouble(blockBuilder, i);
         }
         return new Block[] {blockBuilder.build()};
     }
